removing $Id$ tags in the branch too
svn path=/branches/KDE_3_3_BRANCH/kdelibs/; revision=363701

+++ khtml/html/html_miscimpl.h
@@ -86,7 +86,7 @@ public:
     NodeImpl *nextNamedItem( const DOMString &name ) const;
 
     struct CollectionInfo {
-        CollectionInfo() : version(~0) {}
+        CollectionInfo() : version((unsigned int) ~0) {}
         unsigned int version;
         NodeImpl *current;
         unsigned int position;
